issue #147 refactor: change landing components's directory
